requirements.txt
# Core Dependencies
solana==0.30.2
anchorpy==0.17.0
solders==0.18.1
base58==2.1.1
construct==2.10.68
rich==13.6.0

# Networking
aiohttp==3.8.5
websockets==11.0.3
requests==2.31.0

# Data Processing
pandas==2.1.1
numpy==1.24.3

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# UI & Monitoring
rich==13.6.0
tabulate==0.9.0
plotly==5.17.0

# System
setuptools>=65.5.1
psutil==5.9.5

# Security
cryptography==41.0.4

# Error Handling
tenacity==8.2.3